+++ backend/packages/gates/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @cat-factory/gates
 
+## 0.9.23
+
+### Patch Changes
+
+- Updated dependencies [24f76f1]
+- Updated dependencies [964cfa6]
+  - @cat-factory/contracts@0.261.0
+  - @cat-factory/kernel@0.259.0
+
+## 0.9.22
+
+### Patch Changes
+
+- Updated dependencies [ae44914]
+- Updated dependencies [4be3510]
+  - @cat-factory/contracts@0.260.0
+  - @cat-factory/kernel@0.258.0
+
+## 0.9.21
+
+### Patch Changes
+
+- Updated dependencies [11dae5b]
+  - @cat-factory/contracts@0.259.0
+  - @cat-factory/kernel@0.257.0
+
 ## 0.9.20
 
 ### Patch Changes